fix(app): vue demo new - separate child components
